Rename findr to findr.sh

+++ zsh/zsh_install.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-sudo apt update
-sudo apt install -y zsh zsh-autosuggestions zsh-syntax-highlighting
-sudo apt install -y exa
+if [ -f /usr/bin/apt ]; then
+  sudo apt update
+  sudo apt install -y zsh zsh-autosuggestions zsh-syntax-highlighting
+  sudo apt install -y exa
+else
+  sudo yum install -y zsh zsh-autosuggestions zsh-syntax-highlighting
+fi
 
 sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
